refactored Payments component
