+++ src/applications/disability-benefits/view-payments/components/view-payments-lists/helpers.js
@@ -33,3 +33,25 @@ export const clientServerErrorContent = receivedOrReturned => (
     <p>We were unable to get {receivedOrReturned} payments for your account.</p>
   </>
 );
+
+export const paymentsRecievedContent = (
+  <>
+    <h3 className="vads-u-font-size--xl">Payments you received</h3>
+    <p>
+      VA pays benefits on the first day of the month for the previous month. If
+      the first day of the month is a weekend or holiday, VA pays benefits on
+      the last business day before the 1st. For example, if May 1st is a
+      Saturday, benefits would be paid on Friday, April 30.
+    </p>
+  </>
+);
+
+export const paymentsReturnedContent = (
+  <>
+    <h3 className="vads-u-font-size--xl">Payments returned</h3>
+    <p>
+      Returned payment information is available for 6 years from the date the
+      payment was issued.
+    </p>
+  </>
+);
